Ajustando CartController para validar alteração do coupon

+++ backend/GeekShopping/GeekShopping.CartAPI/Data/ValueObjects/CouponVO.cs
@@ -4,6 +4,6 @@
     {
         public long Id { get; set; }
         public string CouponCode { get; set; }
-        public float DiscountAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
     }
 }
